fix tests not working
